deleting books, finding libraries by book_id

+++ src/main/java/com/example/demo/entities/Library.java
@@ -11,15 +11,15 @@ public class Library {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String name;
 
     @ManyToMany
     @JoinTable(
-            name = "library_book", // The name of the join table
-            joinColumns = @JoinColumn(name = "library_id"), // Foreign key column for Library
-            inverseJoinColumns = @JoinColumn(name = "book_id") // Foreign key column for Book
+            name = "library_book",
+            joinColumns = @JoinColumn(name = "library_id"),
+            inverseJoinColumns = @JoinColumn(name = "book_id")
     )
     private List<Book> books = new ArrayList<>();
 
@@ -29,11 +29,11 @@ public class Library {
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
